snapstate: move config defaults tests to a separate file.
